Refactor mocking fetch to a utility file
