+++ hrm-form-definitions/src/__tests__/formDefinition/loadDefinition.test.ts
@@ -16,82 +16,20 @@
 
 import { DefinitionVersionId, loadDefinition } from '../../formDefinition';
 
-global.fetch = jest.fn();
-const fetchSpy = jest.spyOn(global, 'fetch');
+import { useFetchDefinitions } from '../fetchDefinitionsMock';
 
-const BASE_URL = 'base-url';
-
-const files = [
-  'LayoutDefinitions.json',
-  'caseForms/HouseholdForm.json',
-  'caseForms/IncidentForm.json',
-  'caseForms/NoteForm.json',
-  'caseForms/PerpetratorForm.json',
-  'caseForms/ReferralForm.json',
-  'caseForms/DocumentForm.json',
-  'tabbedForms/CallerInformationTab.json',
-  'tabbedForms/CaseInformationTab.json',
-  'tabbedForms/ChildInformationTab.json',
-  'tabbedForms/IssueCategorizationTab.json',
-  'tabbedForms/ContactlessTaskTab.json',
-  'CallTypeButtons.json',
-  'HelplineInformation.json',
-  'CannedResponses.json',
-  'insights/oneToOneConfigSpec.json',
-  'insights/oneToManyConfigSpecs.json',
-  'CaseStatus.json',
-  'PrepopulateKeys.json',
-  'ReferenceData.json',
-  'BlockedEmojis.json',
-];
-
-const loadJSON = async (jsonPath: string) => {
-  try {
-    // eslint-disable-next-line import/no-dynamic-require, global-require
-    return require(jsonPath);
-  } catch {
-    return undefined;
-  }
-};
-
-// TODO: Refactor to make this more elegant
-const mockFetchImplementation = async (
-  formDefinitionsBaseUrl: string,
-  definitionVersionId: DefinitionVersionId,
-) => {
-  const map = {};
-  for (let i = 0; i < files.length; i += 1) {
-    const file = files[i];
-    const remotePath = `${formDefinitionsBaseUrl}/${file}`;
-    const localPath = `../../../form-definitions/${definitionVersionId}/${file}`;
-
-    // eslint-disable-next-line no-await-in-loop
-    const jsonFile = await loadJSON(localPath);
-
-    const fileExists = !!jsonFile;
-    const mockResponse = {
-      ok: fileExists,
-      status: fileExists ? 200 : 404,
-      json: () => Promise.resolve(jsonFile),
-    };
-
-    map[remotePath] = mockResponse;
-  }
-
-  fetchSpy.mockImplementation((url) => Promise.resolve(map[url.toString()]));
-};
+const { mockFetchImplementation, mockReset, buildBaseURL } = useFetchDefinitions();
 
 beforeEach(() => {
-  fetchSpy.mockReset();
+  mockReset();
 });
 
 describe('loadDefinition', () => {
   test.each(Object.values(DefinitionVersionId))(
     '%p - successfully loads basic structure',
     async (definitionVersionId: DefinitionVersionId) => {
-      const formDefinitionsBaseUrl = `${BASE_URL}/${definitionVersionId}`;
-
-      await mockFetchImplementation(formDefinitionsBaseUrl, definitionVersionId);
+      const formDefinitionsBaseUrl = buildBaseURL(definitionVersionId);
+      await mockFetchImplementation(formDefinitionsBaseUrl);
 
       const definitions = await loadDefinition(formDefinitionsBaseUrl);
       expect(definitions.cannedResponses).toBeInstanceOf(Array);
